data: bolt.service: Disable systemd watchdog to mitigate SIGABRT
Many issues show boltd is frequently terminated by SIGABRT at random
addresses. Boltd relies on the systemd watchdog [1] to report its health.
If this watchdog times out, systemd kills the boltd process. We started to
observe if this change could mitigate this issue.

[1] https://www.freedesktop.org/software/systemd/man/latest/systemd.service.html

...
WatchdogSec=
...
If the time between two such calls is larger than the configured time,
then the service is placed in a failed state and it will be terminated
with SIGABRT
...