boltd: bolt-store: Reset version when failing to get version value
The config file "/var/lib/boltd/version" may contain an empty string. That led
bolt exiting with an error message.

boltd[1906]: could not create manager: failed to parse '' as unsigned integer

This commit will reset the version file and assume the file store version is
set as the default version when failing to get the version value.

Resolves: #194