tests: test-integration: Fix test failure due to the change of pygobject 3.52.3
The recent pygobject release changed the Gio.DBusError from a class
instance to an enum so we can't invoke
Gio.DBusError.ACCESS_DEBIED.quark() to get the quark number. The changes
in this commit get the domain and code from the enum directly.

Moreover, the implementation of Gio.IOErrorEnum and Gio.DBusError are
different. This commit also processes both enums in two separate functions.

Resolves: #195